# For performance reason, don't try to parse XML value in MatchesJsonPathPattern