--- a/src/main/java/com/github/tomakehurst/wiremock/matching/MatchesJsonPathPattern.java
+++ b/src/main/java/com/github/tomakehurst/wiremock/matching/MatchesJsonPathPattern.java
@@ -43,6 +43,13 @@ public class MatchesJsonPathPattern extends PathPattern {
     }
 
     protected MatchResult isSimpleMatch(String value) {
+        // For performance reason, don't try to parse XML value
+        if (value != null && value.trim().startsWith("<")) {
+            notifier().info(String.format(
+                    "Warning: JSON path expression '%s' failed to match document '%s' because it's not JSON document",
+                    expectedValue, value));
+            return MatchResult.noMatch();
+        }
         try {
             Object obj = JsonPath.read(value, expectedValue);
 
@@ -77,6 +84,13 @@ public class MatchesJsonPathPattern extends PathPattern {
     }
 
     protected MatchResult isAdvancedMatch(String value) {
+        // For performance reason, don't try to parse XML value
+        if (value != null && value.trim().startsWith("<")) {
+            notifier().info(String.format(
+                    "Warning: JSON path expression '%s' failed to match document '%s' because it's not JSON document",
+                    expectedValue, value));
+            return MatchResult.noMatch();
+        }
         Object obj = null;
         try {
             obj = JsonPath.read(value, expectedValue);