We have a mix of XML and JSON stubs, and Wiremock try to parse XML request body in JSON.
So, if request body starts with a '<', don't try to parse it in JSON

**With this simple test case:**
```java
@Test
public void perfJsonTest() {
	perfRunJsonPath("matchingJsonPath on XML datas", " <perf-test>perf-data</perf-test> ");
	perfRunJsonPath("matchingJsonPath on JSON datas", " {\"perf-test\": \"perf-data\"} ");
}
private void perfRunJsonPath(String message, String value) {
	long ts = System.currentTimeMillis();
	for (int i = 0; i < 100000; i++) {
		WireMock.matchingJsonPath("$.perf-test").match(value);
	}
	System.out.println(message + ":\t" + (System.currentTimeMillis() - ts) + "ms");
}
```
**Without my merge request:**
matchingJsonPath on XML datas:	1411ms
matchingJsonPath on JSON datas:	487ms

**With my merge request:**
matchingJsonPath on XML datas:	291ms
matchingJsonPath on JSON datas:	481ms